Update README.md
Project 1: A code to map and show the location of fault line movement and earthquake.
Project 2:  A code to show the increase in temperature in September 2002.
Project 3: A code to map out stars and sun(s) in a star cluster.
Project 4: A code to see how similar the DNA of three organisms were, and show the results.
Project 5: Was a code to find out the number of moccasin patters were most and least common and the colors involved.

For all project we used the pandas and matplolib libraries to aid us in finding and coding the results, we also used numpy for many but not all projects. Project 2, 3, and 4 used the numpy library.